requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pyserial==3.5
websockets==12.0
